SMV: fix problem with slice file display on 64 bit Linux and Windows systems -
Former-commit-id: 45bb6fdaf203dbea9ef90e89a74c8eb346d828ff [formerly 0b493a07dacd5d3593222738d773ecf53fb5c6ea]
Former-commit-id: 5d392f1e94548b092c4546ca6648124643cb139b
Former-commit-id: 090b216c746b116a39426eb044ec581a41110551

+++ SMV_5/source/smokeview/IOslice.c
@@ -251,7 +251,7 @@ void readvslice(int ivslice, int flag, int *errorcode){
 /* ------------------ readslice ------------------------ */
 
 void readslice(char *file, int ifile, int flag, int *errorcode){
-  size_t slicefilelen;
+  unsigned int slicefilelen;
   float *xplt_local, *yplt_local, *zplt_local;
   int blocknumber;
   int error;
@@ -1170,7 +1170,7 @@ void getsliceparams(void){
   char *file;
   slice *sd,*sdold;
   int error;
-  size_t lenfile;
+  unsigned int  lenfile;
   float position;
   int is1, is2, js1, js2, ks1, ks2;
   int iblock;
